crud pet, crud business

+++ app/main/service/breed_service.py
@@ -4,7 +4,7 @@ import datetime
 from app.main import db
 from app.main.model.breed import Breed
 from app.main.model.specie import Specie
-
+from app.main.model.pet import Pet
 
 def save_new_breed(data):
     breed = Breed.query.filter_by(name=data['name']).first()
@@ -52,9 +52,11 @@ def patch_a_breed(public_id, data):
 def delete_a_breed(public_id, data):
     breed = Breed.query.filter_by(public_id=public_id).first()
     specie = Specie.query.filter_by(public_id=data["parent_id"]).first()
+    pet = Pet.query.filter_by(breed_subgroup_id=breed.public_id).first()
+
     if breed and specie:
         if breed.specie_parent_id==specie.public_id:
-            if data["name"] == breed.name:
+            if data["name"] == breed.name and not pet:
                 db.session.delete(breed)
                 db.session.commit()
                 response_object = {
@@ -62,6 +64,12 @@ def delete_a_breed(public_id, data):
                     'message': 'Breed successfully deleted.'
                 }
                 return response_object, 201
+            elif pet:
+                response_object = {
+                    'status': 'fail',
+                    'message': 'User created pets depend on this breed.'
+                }
+                return response_object, 405
             else:
                 response_object = {
                     'status': 'fail',
@@ -125,7 +133,29 @@ def get_a_breed(public_id):
     }, 404
 
 def get_all_by_specie(specie_id):
-    return Breed.query.filter_by(specie_parent_id=specie_id).all()
+    breeds = [
+        dict(
+            public_id = breed[0],
+            name = breed[1],
+            parent_id = breed[2],
+            parent_name = breed[3]
+        ) for breed in db.session.query(
+            Breed.public_id,
+            Breed.name,
+            Specie.public_id,
+            Specie.name
+        ).filter(
+            Breed.specie_parent_id == Specie.public_id
+        ).filter(
+            Specie.public_id == specie_id
+        ).all()
+    ]
+    if breeds:
+        return breeds
+    return {
+        'status': 'fail',
+        'message': 'No breeds found.'
+    }, 404
 
 def save_changes(data):
     db.session.add(data)
